feat(custom parameters): Add new dropdownMenu component, used in moreParameters & header optionsDropdown
